tests/requirements.txt
coverage == 3.5.1
django-nose == 1.0
docutils == 0.8.1
ipython == 0.12
ipdb == 0.6.1
FeinCMS == 1.5.3
Markdown == 2.1.1
nose == 1.1.2
pep8 == 1.0
Sphinx == 1.1.3
textile == 2.1.5
